+++ src/lib.rs
@@ -1,7 +1,9 @@
 pub mod contract;
 mod error;
 pub mod ibc;
+pub mod ibc_utils;
 pub mod msg;
 pub mod state;
+pub mod utils;
 
 pub use crate::error::ContractError;
